Refactor: losses module
