+++ .pre-commit-config.yaml
@@ -30,7 +30,7 @@ repos:
     hooks:
       - id: mypy
         files: "^src/"
-        exclude: "^docs/.*|^src/careamics/lvae_training/.*|^src/careamics/models/lvae/.*|^src/careamics/config/likelihood_model.py|^src/careamics/losses/loss_factory.py|^src/careamics/losses/lvae/losses.py"
+        exclude: "^docs/.*|^src/careamics/lvae_training/.*|^src/careamics/models/lvae/.*|^src/careamics/losses/lvae/.*|^src/careamics/config/likelihood_model.py"
         args: ["--config-file", "mypy.ini"]
         additional_dependencies:
           - numpy
