+++ src/careamics/compat/lightning/lightning_module.py
@@ -8,6 +8,7 @@ import torch
 
 from careamics.compat.config.configuration_factories import algorithm_factory
 from careamics.compat.config.data.tile_information import TileInformation
+from careamics.compat.losses import loss_factory
 from careamics.compat.transforms.normalize import Denormalize, TrainDenormalize
 from careamics.compat.transforms.tta import ImageRestorationTTA
 from careamics.config import (
@@ -22,7 +23,6 @@ from careamics.config.support import (
     SupportedOptimizer,
     SupportedScheduler,
 )
-from careamics.losses import loss_factory
 from careamics.models.lvae.noise_models import (
     GaussianMixtureNoiseModel,
     MultiChannelNoiseModel,
