+++ src/careamics/compat/losses/__init__.py
@@ -1,0 +1,7 @@
+"""Losses module."""
+
+__all__ = [
+    "loss_factory",
+]
+
+from .loss_factory import loss_factory
