+++ src/careamics/losses/__init__.py
@@ -4,18 +4,16 @@ __all__ = [
     "denoisplit_loss",
     "denoisplit_musplit_loss",
     "hdn_loss",
-    "loss_factory",
-    "mae_loss",
-    "mse_loss",
+    "lvae_loss_factory",
     "musplit_loss",
     "n2v_loss",
 ]
 
-from .fcn.losses import mae_loss, mse_loss, n2v_loss
-from .loss_factory import loss_factory
-from .lvae.losses import (
+from .lvae import (
     denoisplit_loss,
     denoisplit_musplit_loss,
     hdn_loss,
+    lvae_loss_factory,
     musplit_loss,
 )
+from .n2v_losses import n2v_loss
