--- src/careamics/losses/fcn/__init__.py
@@ -1,1 +1,0 @@
-"""FCN losses."""
