+++ src/careamics/losses/lvae/__init__.py
@@ -1,1 +1,18 @@
 """LVAE losses."""
+
+__all__ = [
+    "denoisplit_loss",
+    "denoisplit_musplit_loss",
+    "hdn_loss",
+    "lvae_loss_factory",
+    "musplit_loss",
+    "n2v_loss",
+]
+
+from .lvae_loss_factory import lvae_loss_factory
+from .lvae_losses import (
+    denoisplit_loss,
+    denoisplit_musplit_loss,
+    hdn_loss,
+    musplit_loss,
+)
